pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "blood-vessel-segmentation"
description = "Package"
readme = "README.md"
requires-python = ">= 3.10"
version = "0.0.1"
dependencies = [
    "lightning~=2.1.3",
    "numexpr==2.8.4",  # due to bug: https://github.com/pandas-dev/pandas/issues/54449
    "monai~=1.3.0",
    "segmentation-models-pytorch~=0.3.3",
#    "segmentation_models_pytorch==0.3.3",
#
    "nptyping~=2.5.0",
    "pydantic~=2.5.3"
#    "scikit-learn==1.3.1",
#    "polars~=0.19.15",

#    "numpy_utility@git+https://github.com/yomura-yomura/numpy_utility#egg=88e5bea"
]

[project.optional-dependencies]
dev = [
    "build~=1.0.3",
    "kaggle~=1.5.16",

    "numpy_utility@git+https://github.com/yomura-yomura/numpy_utility#egg=88e5bea",

    "pandas~=2.1.1",
#    "pyarrow~=13.0.0",
#    "fastparquet~=2023.8.0",
#    "scikit-learn~=1.3.1",
    "wandb~=0.15.12",
#    "toml~=0.10.2",
    "mypy~=1.8.0",
    "pytest~=7.4.3",
    "pytest-xdist~=3.5.0",

    "scipy~=1.11.4",
    "numba~=0.58.1",
    "scikit-image~=0.22.0"
]
dash = [
    "dash[diskcache]~=2.12.1",
    "dash_bootstrap_components~=1.4.1",
    "dash_auth~=1.4.1"
]
pycharm = [
    "ipython ~= 8.19.0",
    "plotly_utility@git+https://github.com/yomura-yomura/plotly_utility#egg=10757e5d1f9b44dc6bab19bd4970b8d9f09bb973",
    "kaleido ~= 0.2.1"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "blood_vessel_segmentation"
]

[tool.isort]
include_trailing_comma = true
multi_line_output = 3
line_length = 120
profile = "black"

[tool.black]
line-length = 120

[tool.mypy]
show_error_context = true
strict = true
show_column_numbers = true
namespace_packages = true
disallow_subclassing_any = false
disallow_untyped_decorators = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
follow_imports = "silent"
exclude = []
